Split off LoadSampleDatasetCard
